[refactor] テスト用ヘルパのファイル名を _helpers.py から utils.py に変更 (refs #127 )
